Added @Par annotation and unit tests for parameters (without implementation yet)
