+++ src/main/java/ru/redenergy/flexy/annotation/Par.java
@@ -1,0 +1,15 @@
+package ru.redenergy.flexy.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * Optional parameter of the command
+ */
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Par {
+    String value();
+}
